+++ app/Http/Controllers/Admin/MovieController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Events\MovieCreated;
+use App\Jobs\AuditLogJob;
 use App\Models\Movie;
 use App\Models\Genre;
 use App\Models\Director;
 use App\Models\Actor;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use App\Http\Requests\MovieRequest;
@@ -16,16 +18,16 @@ use App\Jobs\ProcessPosterImage;
 
 class MovieController extends Controller
 {
-
-
     public function index()
     {
+        Gate::authorize('viewAny', Movie::class);
         $movies = Movie::with(['genres', 'director', 'actors'])->get();
         return view('admin.movies.index', compact('movies'));
     }
 
     public function create()
     {
+        Gate::authorize('create', Movie::class);
         $genres = Genre::all();
         $directors = Director::all();
         $actors = Actor::all();
@@ -34,17 +36,17 @@ class MovieController extends Controller
 
     public function store(MovieRequest $request)
     {
+        Gate::authorize('create', Movie::class);
         $data = $request->validated();
 
 
-        // 2. Procesamos el póster y lo metemos en el array de datos
+        // 2. Procesamos el póster
         if ($request->hasFile('poster')) {
             $path = $request->file('poster')->store('posters', 'public');
-            $data['poster'] = $path; // <-- Lo guardamos en $data, no en una variable suelta
+            $data['poster'] = $path;
         }
 
-        // 3. Pasamos TODO el array $data a create.
-        // Aquí es donde irán title, year, description, duration, etc.
+        // 3. Crear
         $movie = Movie::create($data);
 
         // 4. Relaciones
@@ -53,9 +55,8 @@ class MovieController extends Controller
             $movie->actors()->sync($request->actors);
         }
 
-        // 5. Tus procesos extra
-        \App\Jobs\ProcessPosterImage::dispatch($movie->title);
-        \App\Events\MovieCreated::dispatch($movie);
+        // Job
+        AuditLogJob::dispatch("PELÍCULA: Se ha creado la película '{$movie->title}' por " . Auth::user()->email);;
         \Illuminate\Support\Facades\Artisan::call('movies:clean');
 
         return redirect()->route('movies.index')
@@ -64,12 +65,14 @@ class MovieController extends Controller
 
     public function show(Movie $movie)
     {
+        Gate::authorize('view', $movie);
         $movie->load(['genres', 'director', 'actors']);
         return view('admin.movies.show', compact('movie'));
     }
 
     public function edit(Movie $movie)
     {
+        Gate::authorize('update', $movie);
         $genres = Genre::all();
         $directors = Director::all();
         $actors = Actor::all();
@@ -80,6 +83,7 @@ class MovieController extends Controller
 
     public function update(MovieRequest $request, Movie $movie)
     {
+        Gate::authorize('update', $movie);
         $data = $request->validated();
 
         if ($request->hasFile('poster')) {
@@ -89,13 +93,14 @@ class MovieController extends Controller
             $data['poster'] = $request->file('poster')->store('posters', 'public');
         }
 
-        // si cambio el director_id se actualiza
         $movie->update($data);
 
-        // Sincronizamos generos y actores
-
+        // Sincronizamos
         $movie->genres()->sync($request->input('genres', []));
         $movie->actors()->sync($request->input('actors', []));
+
+        //Job
+        AuditLogJob::dispatch("PELÍCULA: Se ha editado la película '{$movie->title}' por " . Auth::user()->email);
 
         return redirect()->route('movies.index')
             ->with('success', 'Película actualizada correctamente.');
@@ -103,11 +108,18 @@ class MovieController extends Controller
 
     public function destroy(Movie $movie)
     {
-        // Gate::authorize('delete', $movie);
+        Gate::authorize('delete', $movie); // Solo Admin
+
+        $title = $movie->title;
+
         if ($movie->poster) {
             Storage::disk('public')->delete($movie->poster);
         }
+
         $movie->delete();
+
+        AuditLogJob::dispatch("PELÍCULA: Se ha eliminado la película '$title' por " . Auth::user()->email);
+
         return redirect()->route('movies.index')->with('success', 'Película eliminada correctamente.');
     }
 }
